Restaurar

+++ app/Http/Controllers/Tb_materia_prima_productoController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Tb_materia_prima_producto;
 use App\Tb_gestion_materia_prima;
 use App\Tb_unidad_base;
-use App\Tb_kardex_almacen;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -78,8 +77,7 @@ class Tb_materia_prima_productoController extends Controller
 
     }
 
-    public function selectMateriaPrimaProducto()
-    {
+    public function selectMateriaPrimaProducto(){
         $materiaprimaproductos = Tb_materia_prima_producto::join("tb_gestion_materia_prima","tb_materia_prima_producto.idMateriaPrima","=","tb_gestion_materia_prima.id")
         ->leftJoin('tb_unidad_base',function($join){
             $join->on('tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id');
@@ -103,7 +101,7 @@ class Tb_materia_prima_productoController extends Controller
                 'materiaprimaproductos' => $materiaprimaproductos
         ];
 
-    }
+        }
 
         public function store(Request $request)
         {
@@ -129,16 +127,15 @@ class Tb_materia_prima_productoController extends Controller
 
         public function deactivate(Request $request)
         {
-            //if(!$request->ajax()) return redirect('/');
+            if(!$request->ajax()) return redirect('/');
             DB::statement('SET FOREIGN_KEY_CHECKS=0;');
             $tb_materia_prima_producto=Tb_materia_prima_producto::findOrFail($request->id);
             $tb_materia_prima_producto->delete();
-            DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+           DB::statement('SET FOREIGN_KEY_CHECKS=1;');
         }
 
 
-        public function selectGestionMateria()
-        {
+        public function selectGestionMateria(){
             $gestionmaterias = Tb_gestion_materia_prima::join('tb_unidad_base','tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id')
             ->select('tb_gestion_materia_prima.id as idGestionMateria','tb_gestion_materia_prima.gestionMateria','tb_gestion_materia_prima.precioBase','tb_gestion_materia_prima.idUnidadBase','tb_unidad_base.unidadBase as unidadBase','tb_gestion_materia_prima.estado')
             ->where('tb_gestion_materia_prima.estado','=','1')
@@ -148,8 +145,7 @@ class Tb_materia_prima_productoController extends Controller
             return ['gestionmaterias' => $gestionmaterias];
         }
 
-        public function selectDatosMateria($id)
-        {
+        public function selectDatosMateria($id){
             $datosmaterias = Tb_gestion_materia_prima::join('tb_unidad_base','tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id')
             ->select('tb_gestion_materia_prima.id as idGestion','tb_gestion_materia_prima.gestionMateria as nombreMateria','tb_gestion_materia_prima.precioBase as precioBase','tb_unidad_base.unidadBase as unidadBase')
             ->where('tb_gestion_materia_prima.id','=',$id)
@@ -158,61 +154,7 @@ class Tb_materia_prima_productoController extends Controller
             return ['datosmaterias' => $datosmaterias];
         }
 
-        public function valorPrecioBase($id)
-        { //DATOS de valor segun orden 2 5 y 6 traigo segun idmateria el valor promedio del kardex
-            //if(!$request->ajax()) return redirect('/');
-            $valorMaterial = 0;
-            $preciomaterial = Tb_kardex_almacen::first()
-            ->select('tb_kardex_almacen.id','tb_kardex_almacen.precioSaldos')
-            ->where([
-                ['tb_kardex_almacen.idGestionMateria', '=', $id],
-                ['tb_kardex_almacen.precioSaldos', '>', '0'],
-            ]) //aca podria poner en el where que mirara que el valor no es cero y probar edit funciona
-            ->orderBy('tb_kardex_almacen.idGestionMateria','desc')
-            ->get();
-
-            foreach($preciomaterial as $totalg){
-                $id = $totalg->id;
-                $valorMaterial = $valorMaterial + $totalg->precioSaldos;
-            }
-
-            if($valorMaterial>0){
-                $valores = Tb_gestion_materia_prima::join('tb_unidad_base','tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id')
-                ->select('tb_unidad_base.unidadBase as unidadBase')
-                ->where('tb_gestion_materia_prima.id','=',$id)
-                ->get();
-
-                foreach($valores as $totalg){
-                    $unidad = $totalg->unidadBase;
-                }
-
-                $valorPrecioBase = $valorMaterial;
-            }
-            else{
-                $precioB = 0;
-                $valores = Tb_gestion_materia_prima::join('tb_unidad_base','tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id')
-                ->select('tb_gestion_materia_prima.precioBase as precioBase','tb_unidad_base.unidadBase as unidadBase')
-                ->where('tb_gestion_materia_prima.id','=',$id)
-                ->get();
-
-                foreach($valores as $totalg){
-                    $valor = $totalg->precioBase + $precioB;
-                    $unidad = $totalg->unidadBase;
-                }
-
-                $valorPrecioBase = $valor;
-            }
-
-            return [
-            'valorPrecioBase' => $valorPrecioBase,
-            'unidadBase' => $unidad
-            ];
-
-        }
-
-/*
-        public function valorPrecioBase($id)
-        {
+        public function valorPrecioBase($id){
             $precioB = 0;
             $valores = Tb_gestion_materia_prima::join('tb_unidad_base','tb_gestion_materia_prima.idUnidadBase','=','tb_unidad_base.id')
             ->select('tb_gestion_materia_prima.precioBase as precioBase','tb_unidad_base.unidadBase as unidadBase')
@@ -228,6 +170,23 @@ class Tb_materia_prima_productoController extends Controller
             return ['valorPrecioBase' => $valorPrecioBase,
                     'unidadBase' => $unidad
                     ];
+        }
+/*
+        public function valorPrecioBase($id){
+            $precioB = 0;
+
+            //valorMinuto Perfil
+            $valores = tb_gestion_materia_prima::where([
+                ['id','=',$id],
+            ])
+            ->select('precioBase')->get();
+
+            foreach($valores as $totalg){
+                $valor = $totalg->precioBase + $precioB;
+            }
+
+            $valorPrecioBase = $valor;
+            return ['valorPrecioBase' => $valorPrecioBase];
         }
 */
